protect against missing setup.  More cleanup
